requirements.txt
Pillow==2.8.1
backports.ssl-match-hostname==3.4.0.2
certifi==2015.04.28
tornado==4.1
wsgiref==0.1.2
pyserial
adafruit_bbio
